+++ R/spada_ui.R
@@ -7,11 +7,10 @@ spada_ui <- function(){
     busy_start_up(
       loader = spin_epic('orbit', color = '#FFFFFF'),
       text = tagList(
-        h1(
-        'Spada',
+        h1('Spada',
         style = "font-family: 'Times'; font-size: 120px;"),
-        p('Loading...',
-          style = "font-family: 'Times'; font-size: 20px;")
+        h3('R > Shiny > You',
+           style = "font-family: 'Times'; font-size: 20px;")
       ),
       mode = 'auto',
       timeout = 1200,